Move SafeLineEnvVarLoader to its own file.
