+++ earthmover/yaml_parser.py
@@ -1,0 +1,23 @@
+import os
+
+from yaml import SafeLoader
+
+
+# This allows us to determine the YAML file line number for any element loaded from YAML
+# (very useful for debugging and giving meaningful error messages)
+# (derived from https://stackoverflow.com/a/53647080)
+# Also added env var interpolation based on
+# https://stackoverflow.com/questions/52412297/how-to-replace-environment-variable-value-in-yaml-file-to-be-parsed-using-python#answer-55301129
+class SafeLineEnvVarLoader(SafeLoader):
+
+    def construct_mapping(self, node, deep=False):
+        mapping = super().construct_mapping(node, deep=deep)
+
+        # expand env vars:
+        for k, v in mapping.items():
+            if isinstance(v, str):
+                mapping[k] = os.path.expandvars(v)
+
+        # Add 1 so line numbering starts at 1
+        mapping['__line__'] = node.start_mark.line + 1
+        return mapping
